Fix argon hash length to make shares valid.
The python argon2_cffi library uses a default hash length of 16 bytes (with a 16 byte salt).  This conflicts with the default in the PHP reference client (and the code the pool uses) which is a 32 byte hash with 16 byte salt.  

As such, shares made from the client would be considered invalid thusfar, resulting in the pool blocking further shares from the workers running this client.  This commit changes the parameters passed to the argon2 'PasswordHasher' constructor to set the length correctly to 32 bytes.